Remove ReleaseReadinessTool and move it into SdkReleaseTool
