+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Tools/Package/SdkReleaseTool.cs
@@ -1,6 +1,7 @@
 using System.CommandLine;
 using System.CommandLine.Parsing;
 using System.ComponentModel;
+using Microsoft.TeamFoundation.Build.WebApi;
 using ModelContextProtocol.Server;
 using Azure.Sdk.Tools.Cli.Commands;
 using Azure.Sdk.Tools.Cli.Models;
@@ -10,9 +11,10 @@ using Azure.Sdk.Tools.Cli.Models.Responses.Package;
 namespace Azure.Sdk.Tools.Cli.Tools.Package
 {
     [McpServerToolType, Description("This type contains the tools to release SDK package")]
-    public class SdkReleaseTool(IDevOpsService devopsService, ILogger<SdkReleaseTool> logger, ILogger<ReleaseReadinessTool> releaseReadinessLogger) : MCPTool
+    public class SdkReleaseTool(IDevOpsService devopsService, ILogger<SdkReleaseTool> logger) : MCPTool
     {
         private const string ReleaseSdkToolName = "azsdk_release_sdk";
+        private const string Pipeline_Success_Status = "Succeeded";
         
         public override CommandGroup[] CommandHierarchy { get; set; } = [SharedCommandGroups.Package];
 
@@ -59,7 +61,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.Package
         }
 
         [McpServerTool(Name = ReleaseSdkToolName), Description("Releases the specified SDK package for a language. This includes checking if the package is ready for release and triggering the release pipeline. To ONLY check package release readiness pass dryRun as true.")]
-        public async Task<SdkReleaseResponse> ReleasePackageAsync(string packageName, string language, string branch = "main")
+        public async Task<SdkReleaseResponse> ReleasePackageAsync(string packageName, string language, string branch = "main", bool dryRun = false)
         {
             try
             {
@@ -105,8 +107,7 @@ namespace Azure.Sdk.Tools.Cli.Tools.Package
                 }
 
                 // Check if the package is ready for release
-                var releaseReadinessTool = new ReleaseReadinessTool(devopsService, releaseReadinessLogger);
-                var releaseReadiness = await releaseReadinessTool.CheckPackageReleaseReadinessAsync(packageName, language);
+                var releaseReadiness = await CheckPackageReleaseReadinessAsync(packageName, language);
                 if (!releaseReadiness.IsPackageReady)
                 {
                     response.ReleaseStatusDetails = $"Package is not ready for release. {releaseReadiness.PackageReadinessDetails}";
@@ -167,6 +168,135 @@ namespace Azure.Sdk.Tools.Cli.Tools.Package
                 };
                 response.SetLanguage(language);
                 return response;
+            }
+        }
+
+        private async Task<PackageWorkitemResponse> CheckPackageReleaseReadinessAsync(string packageName, string language)
+        {
+            try
+            {
+                var package = await devopsService.GetPackageWorkItemAsync(packageName, language);
+                if (package == null)
+                {
+                    package = new PackageWorkitemResponse
+                    {
+                        PackageName = packageName,   
+                        ResponseError = $"No package work item found for package '{packageName}' in language '{language}'. Please check the package name and language."
+                    };
+                    package.SetLanguage(language);
+                    return package;
+                }
+
+                package.IsPackageReady = package.IsChangeLogReady;
+
+                //Check release date for latest version in planned release
+                var plannedRelease = package.PlannedReleases.FirstOrDefault(r => r.Version.Equals(package.Version)) ?? package.PlannedReleases.LastOrDefault();
+                package.PlannedReleaseDate = plannedRelease?.ReleaseDate ?? "Unknown";
+                if (package.PlannedReleaseDate.Equals("Unknown"))
+                {
+                    package.IsPackageReady = false;
+                    package.PackageReadinessDetails = $"No planned release date found in package details for current package version {package.Version}. Please check the package version and verify that change log file is correct. ";
+                }
+
+                var releaseType = plannedRelease?.ReleaseType ?? "Unknown";
+                bool isPreviewRelease = releaseType.Equals("Beta");
+                bool isDataPlanePackage = package.PackageType != SdkType.Management;
+                // Check for namespace approval if preview release for data plane
+                if (isDataPlanePackage && isPreviewRelease)
+                {
+                    if (!package.IsPackageNameApproved)
+                    {
+                        package.IsPackageReady = false;
+                        package.PackageReadinessDetails += $"Package name '{packageName}' is not approved for preview release. ";
+                    }
+                    // no need to add extra package name approval status if package name is approved or has at least one version already released
+                }
+                else
+                {
+                    package.PackageNameStatus = "Not required";
+                    package.PackageNameApprovalDetails = "Package name approval is not required for GA releases of data plane packages or for non-data plane packages.";
+                }
+
+                // Check if API view is approved if stable version for data plane or .NET
+                if ((isDataPlanePackage || language.Equals(".NET")) && !isPreviewRelease)
+                {
+
+                    if (!package.IsApiViewApproved)
+                    {
+                        package.IsPackageReady = false;
+                        package.PackageReadinessDetails += $"API view is not approved for GA release of package '{packageName}'. ";
+                    }
+                }
+                else
+                {
+                    package.APIViewStatus = "Not required";
+                    package.ApiViewValidationDetails = "API view is not required for preview releases of data plane packages or for non-data plane packages.";
+                }
+
+                // Check last pipeline run status for the package and verify it completed successfully
+                package.LatestPipelineStatus = await GetPipelineRunDetails(package.LatestPipelineRun);
+                bool hasPipelineWarning = string.IsNullOrEmpty(package.LatestPipelineStatus) || !package.LatestPipelineStatus.Contains(Pipeline_Success_Status);
+
+                // Package release readiness status
+                if (package.IsPackageReady)
+                {
+                    package.PackageReadinessDetails = $"Package '{packageName}' is ready for release. Queue a release pipeline run using the link {package.PipelineDefinitionUrl} to release the package.";
+                    
+                    // Add warning about pipeline status if not successful
+                    if (hasPipelineWarning)
+                    {
+                        package.PackageReadinessDetails += $"\n\nWARNING: The last known CI pipeline status for this package has failed. This might cause issues when running the release pipeline if the error was not a transient failure. Please review the last pipeline run at {package.LatestPipelineRun} to verify the failure was transient before proceeding with the release.";
+                    }
+                }
+                else
+                {
+                    package.PackageReadinessDetails += $"Package '{packageName}' is not ready for release. Please address the issues mentioned above.";
+                }
+                return package;
+            }
+            catch (Exception ex)
+            {
+                var package = new PackageWorkitemResponse
+                {
+                    PackageName = packageName,
+                    IsPackageReady = false,
+                    ResponseError = $"Failed to check package readiness for '{packageName}' in language '{language}'. Error {ex.Message}"
+                };
+                package.SetLanguage(language);
+                return package;
+            }
+        }
+
+        private async Task<string> GetPipelineRunDetails(string pipelineRunUrl)
+        {
+            try
+            {
+                logger.LogInformation("Getting pipeline run details for URL: {pipelineRunUrl}", pipelineRunUrl);
+                if (!string.IsNullOrEmpty(pipelineRunUrl) && pipelineRunUrl.Contains("buildId="))
+                {
+                    var buildId = int.Parse(pipelineRunUrl.Split("buildId=").LastOrDefault());
+                    logger.LogInformation("Extracted build ID: {buildId}", buildId);
+                    var pipelineRun = await devopsService.GetPipelineRunAsync(buildId);
+                    if (pipelineRun != null)
+                    {
+                        logger.LogInformation(
+                            "Pipeline status: {PipelineStatus}, Result: {PipelineResult}",
+                            pipelineRun.Status,
+                            pipelineRun.Result);
+                        var status = (pipelineRun.Status == BuildStatus.Completed ? pipelineRun.Result?.ToString() : pipelineRun.Status.ToString()) ?? "Unknown";
+                        if (!status.Contains(Pipeline_Success_Status))
+                        {
+                            status = $"Pipeline run with ID {buildId} did not succeed. Status: {status}. Please check the pipeline run details at {DevOpsService.GetPipelineUrl(buildId)} for more information.";
+                        }
+                        return status;
+                    }
+                }
+                return $"Failed to get pipeline run details. The pipeline run URL '{pipelineRunUrl}' is invalid or does not contain a valid build ID.";
+            }
+            catch (Exception ex)
+            {
+                logger.LogError(ex, "Failed to get pipeline run details for URL {PipelineRunUrl}", pipelineRunUrl);
+                return $"Failed to get pipeline run details. Error: {ex.Message}";
             }
         }
     }
